[vtflash] constrain_step now return whole amax
Previously, it returns 0.9 of amax. There was a problem, that sometimes
function "descent minimum" lies in [0.9 amax, amax], so lineaserch fails
as second strong Wolfe condition didn't meet.

VTFlash in range of temperature and concentration shows that 0.9999*amax is OK
and prevfloat(amax) is NOT OK. But to avoid ad hoc techniques, now it just return
the edge amax of feasiblity, though amax is NOT FEASIBLE.